refacto: create scanner component
